refactoring props
